Move isAbortSignal to separate file
